legacy/voice_assistant/pyproject.toml
[project]
name = "voice-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.12"
dependencies = [
    "fastapi>=0.121.2",
    "faster-whisper>=1.2.1",
    "groq>=0.35.0",
    "piper-tts>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sounddevice>=0.5.3",
    "soundfile>=0.13.1",
    "uvicorn>=0.38.0",
    "gradio==4.20.1",
    "pyaudio>=0.2.14",
    "huggingface-hub<0.22.0",
    "typer==0.9.0",
]
